Fix theme preview not updating on settings page
The AppProvider (which applies appearance data-attributes to <html>) was
only mounted under /channels. The /settings layout had no equivalent,
so changing themes in appearance settings updated the Zustand store but
never wrote the data-theme-preset attribute, leaving CSS variables stale.

Extract the appearance-application effect into a reusable useApplyAppearance
hook and call it from both AppProvider and a new SettingsAppearanceProvider
wrapper in the settings layout.
